core/requirements.txt
asgiref==3.8.1
attrs==25.1.0
black==25.1.0
click==8.1.8
Django==5.1.6
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.3.1
exceptiongroup==1.2.2
flake8==7.2.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
psycopg==3.2.5
psycopg-binary==3.2.5
pycodestyle==2.13.0
pyflakes==3.3.2
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.11.1
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
